# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shoken WebApp API",
    "description": "証券管理Webアプリ APIドキュメント",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/asset-balances": {
      "get": {
        "tags": [
          "handlers::v1::asset_balances"
        ],
        "summary": "保有銘柄一覧を取得（v1）",
        "operationId": "v1_asset_balance_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetBalance"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "handlers::v1::asset_balances"
        ],
        "summary": "保有銘柄を全置換（v1）\nステータスコード 200 を返す（既存 POST /asset-balances/bulk は 201）",
        "operationId": "v1_asset_balance_replace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateAssetBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "handlers::v1::asset_balances"
        ],
        "summary": "保有銘柄を全削除（v1）",
        "operationId": "v1_asset_balance_delete_all",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/v1/dividends": {
      "get": {
        "tags": [
          "handlers::v1::dividends"
        ],
        "summary": "配当金一覧を取得（v1）",
        "operationId": "v1_dividend_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dividend"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "handlers::v1::dividends"
        ],
        "summary": "配当金を全削除（v1）",
        "operationId": "v1_dividend_delete_all",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/asset-balances": {
      "get": {
        "tags": [
          "handlers::asset_balance"
        ],
        "summary": "認証ユーザーの保有銘柄一覧を取得",
        "operationId": "asset_balance_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetBalance"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "handlers::asset_balance"
        ],
        "summary": "認証ユーザーの保有銘柄を全削除",
        "operationId": "asset_balance_delete_all",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/asset-balances/bulk": {
      "post": {
        "tags": [
          "handlers::asset_balance"
        ],
        "summary": "保有銘柄を一括追加（既存は更新）",
        "operationId": "asset_balance_bulk_create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateAssetBalanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/asset-balances/csv": {
      "post": {
        "tags": [
          "handlers::asset_balance"
        ],
        "summary": "CSV ファイルをアップロードして保有銘柄を一括登録（既存データ全置換）",
        "operationId": "asset_balance_upload_csv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CsvUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/asset-balances/csv/preview": {
      "post": {
        "tags": [
          "handlers::asset_balance"
        ],
        "summary": "CSV ファイルをパースして保存前プレビューを返す（DB 書き込みなし）",
        "operationId": "asset_balance_preview_csv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CsvUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvPreviewResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": [
          "handlers::auth"
        ],
        "summary": "現在ログイン中のユーザー情報を取得",
        "operationId": "auth_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/dividends": {
      "get": {
        "tags": [
          "handlers::dividend"
        ],
        "summary": "認証ユーザーの配当金一覧を取得",
        "operationId": "dividend_list",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dividend"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "handlers::dividend"
        ],
        "summary": "認証ユーザーの配当金を全削除",
        "operationId": "dividend_delete_all",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/dividends/csv": {
      "post": {
        "tags": [
          "handlers::dividend"
        ],
        "summary": "CSV ファイルをアップロードして配当金を一括登録",
        "operationId": "dividend_upload_csv",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CsvUploadForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsvUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetBalance": {
        "type": "object",
        "description": "保有銘柄モデル（DB + APIレスポンス兼用）",
        "required": [
          "id",
          "security_code",
          "security_name",
          "shares",
          "executing_shares",
          "average_purchase_price",
          "total_purchase_amount",
          "current_price",
          "daily_change",
          "market_value",
          "profit_loss_rate",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "average_purchase_price": {
            "type": "number",
            "format": "double"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current_price": {
            "type": "number",
            "format": "double"
          },
          "daily_change": {
            "type": "number",
            "format": "double"
          },
          "executing_shares": {
            "type": "number",
            "format": "double"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "market_value": {
            "type": "number",
            "format": "double"
          },
          "profit_loss_rate": {
            "type": "number",
            "format": "double"
          },
          "security_code": {
            "type": "string"
          },
          "security_name": {
            "type": "string"
          },
          "shares": {
            "type": "number",
            "format": "double"
          },
          "total_purchase_amount": {
            "type": "number",
            "format": "double"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        }
      },
...
      "MessageResponse": {
        "type": "object",
        "description": "メッセージレスポンス（削除・ログアウト等）",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token"
      }
    }
  }
}